Extract EventAggregator interface
