+++ packages/core/src/event-aggregator/gql-service.ts
@@ -1,0 +1,92 @@
+import Emittery from "emittery";
+
+import type { LogFilter } from "@/config/logFilters";
+import type { Network } from "@/config/networks";
+import type { EventStore } from "@/event-store/store";
+import type { Common } from "@/Ponder";
+
+import type {
+  EventAggregatorEvents,
+  EventAggregatorMetrics,
+  EventAggregatorService,
+} from "./service";
+
+export class GqlEventAggregatorService
+  extends Emittery<EventAggregatorEvents>
+  implements EventAggregatorService
+{
+  private networks: Network[];
+
+  // Minimum timestamp at which events are available (across all networks).
+  checkpoint: number;
+  // Minimum finalized timestamp (across all networks).
+  finalityCheckpoint: number;
+
+  // Timestamp at which the historical sync was completed (across all networks).
+  historicalSyncCompletedAt?: number;
+
+  // Per-network event timestamp checkpoints.
+  private networkCheckpoints: Record<
+    number,
+    {
+      isHistoricalSyncComplete: boolean;
+      historicalCheckpoint: number;
+      realtimeCheckpoint: number;
+      finalityCheckpoint: number;
+    }
+  >;
+
+  metrics: EventAggregatorMetrics;
+
+  constructor({
+    networks,
+  }: {
+    common: Common;
+    eventStore: EventStore;
+    networks: Network[];
+    logFilters: LogFilter[];
+  }) {
+    super();
+
+    this.networks = networks;
+    this.metrics = {};
+
+    this.checkpoint = 0;
+    this.finalityCheckpoint = 0;
+
+    this.networkCheckpoints = {};
+    this.networks.forEach((network) => {
+      this.networkCheckpoints[network.chainId] = {
+        isHistoricalSyncComplete: false,
+        historicalCheckpoint: 0,
+        realtimeCheckpoint: 0,
+        finalityCheckpoint: 0,
+      };
+    });
+  }
+
+  /** Fetches events for all registered log filters between the specified timestamps.
+   *
+   * @param options.fromTimestamp Timestamp to start including events (inclusive).
+   * @param options.toTimestamp Timestamp to stop including events (inclusive).
+   * @param options.includeLogFilterEvents Map of log filter name -> selector -> ABI event item for which to include full event objects.
+   * @returns A promise resolving to an array of log events.
+   */
+  async *getEvents() {}
+
+  handleNewHistoricalCheckpoint = () => {};
+
+  handleHistoricalSyncComplete = () => {};
+
+  handleNewRealtimeCheckpoint = () => {};
+
+  handleNewFinalityCheckpoint = () => {};
+
+  handleReorg = ({
+    commonAncestorTimestamp,
+  }: {
+    commonAncestorTimestamp: number;
+  }) => {
+    this.emit("reorg", { commonAncestorTimestamp });
+  };
+}
